requirements.txt
web3
solana
solders
openai
loguru
selenium
webdriver_manager
playwright
aiohttp
quart
dataclasses
python-dotenv
python-telegram-bot
telethon
base58
curl_cffi